+++ src/features/dashboard/model/email.ts
